refactor(web): move ViaVai page to user and remove legacy Form/Table pages
